Introduce useResourceContext
