+++ www/front_src/src/Resources/Actions/Resource/index.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 
+import { isEmpty } from 'ramda';
+
 import { Button, ButtonProps, Grid } from '@material-ui/core';
 import IconAcknowledge from '@material-ui/icons/Person';
 import IconCheck from '@material-ui/icons/Sync';
@@ -14,42 +16,29 @@ import {
   labelSomethingWentWrong,
   labelCheckCommandSent,
 } from '../../translatedLabels';
-import { Resource } from '../../models';
 import AcknowledgeForm from './Acknowledge';
 import DowntimeForm from './Downtime';
 import { checkResources } from '../../api';
-
-interface Props {
-  disabled: boolean;
-  resourcesToAcknowledge: Array<Resource>;
-  onPrepareToAcknowledge;
-  onCancelAcknowledge;
-  resourcesToSetDowntime: Array<Resource>;
-  onPrepareToSetDowntime;
-  onCancelSetDowntime;
-  resourcesToCheck: Array<Resource>;
-  onPrepareToCheck;
-  onSuccess;
-}
+import { useResourceContext } from '../../Context';
 
 const ActionButton = (props: ButtonProps): JSX.Element => (
   <Button variant="contained" color="primary" size="small" {...props} />
 );
 
-const ResourceActions = ({
-  disabled,
-  resourcesToAcknowledge,
-  onPrepareToAcknowledge,
-  onCancelAcknowledge,
-  resourcesToSetDowntime,
-  onPrepareToSetDowntime,
-  onCancelSetDowntime,
-  resourcesToCheck,
-  onPrepareToCheck,
-  onSuccess,
-}: Props & Omit<ButtonProps, 'disabled'>): JSX.Element => {
+const ResourceActions = (): JSX.Element => {
   const { cancel, token } = useCancelTokenSource();
   const { showMessage } = useSnackbar();
+
+  const {
+    resourcesToCheck,
+    setSelectedResources,
+    selectedResources,
+    resourcesToAcknowledge,
+    setResourcesToAcknowledge,
+    resourcesToSetDowntime,
+    setResourcesToSetDowntime,
+    setResourcesToCheck,
+  } = useResourceContext();
 
   const showError = (message): void =>
     showMessage({ message, severity: Severity.error });
@@ -57,6 +46,13 @@ const ResourceActions = ({
     showMessage({ message, severity: Severity.success });
 
   const hasResourcesToCheck = resourcesToCheck.length > 0;
+
+  const confirmAction = (): void => {
+    setSelectedResources([]);
+    setResourcesToAcknowledge([]);
+    setResourcesToSetDowntime([]);
+    setResourcesToCheck([]);
+  };
 
   React.useEffect(() => {
     if (!hasResourcesToCheck) {
@@ -68,13 +64,35 @@ const ResourceActions = ({
       cancelToken: token,
     })
       .then(() => {
+        confirmAction();
         showSuccess(labelCheckCommandSent);
-        onSuccess();
       })
       .catch(() => showError(labelSomethingWentWrong));
   }, [resourcesToCheck]);
 
   React.useEffect(() => (): void => cancel(), []);
+
+  const prepareToAcknowledge = (): void => {
+    setResourcesToAcknowledge(selectedResources);
+  };
+
+  const prepareToSetDowntime = (): void => {
+    setResourcesToSetDowntime(selectedResources);
+  };
+
+  const prepareToCheck = (): void => {
+    setResourcesToCheck(selectedResources);
+  };
+
+  const cancelAcknowledge = (): void => {
+    setResourcesToAcknowledge([]);
+  };
+
+  const cancelSetDowntime = (): void => {
+    setResourcesToSetDowntime([]);
+  };
+
+  const disabled = isEmpty(selectedResources);
 
   return (
     <Grid container spacing={1}>
@@ -82,7 +100,7 @@ const ResourceActions = ({
         <ActionButton
           disabled={disabled}
           startIcon={<IconAcknowledge />}
-          onClick={onPrepareToAcknowledge}
+          onClick={prepareToAcknowledge}
         >
           {labelAcknowledge}
         </ActionButton>
@@ -91,7 +109,7 @@ const ResourceActions = ({
         <ActionButton
           disabled={disabled}
           startIcon={<IconDowntime />}
-          onClick={onPrepareToSetDowntime}
+          onClick={prepareToSetDowntime}
         >
           {labelDowntime}
         </ActionButton>
@@ -100,7 +118,7 @@ const ResourceActions = ({
         <ActionButton
           disabled={disabled}
           startIcon={<IconCheck />}
-          onClick={onPrepareToCheck}
+          onClick={prepareToCheck}
         >
           {labelCheck}
         </ActionButton>
@@ -108,15 +126,15 @@ const ResourceActions = ({
       {resourcesToAcknowledge.length > 0 && (
         <AcknowledgeForm
           resources={resourcesToAcknowledge}
-          onClose={onCancelAcknowledge}
-          onSuccess={onSuccess}
+          onClose={cancelAcknowledge}
+          onSuccess={confirmAction}
         />
       )}
       {resourcesToSetDowntime.length > 0 && (
         <DowntimeForm
           resources={resourcesToSetDowntime}
-          onClose={onCancelSetDowntime}
-          onSuccess={onSuccess}
+          onClose={cancelSetDowntime}
+          onSuccess={confirmAction}
         />
       )}
     </Grid>
