+++ www/front_src/src/Resources/api/index.ts
@@ -1,6 +1,8 @@
-import axios, { AxiosRequestConfig, AxiosResponse, CancelToken } from 'axios';
+import axios, { AxiosResponse, CancelToken } from 'axios';
 import formatISO from 'date-fns/formatISO';
 import { map, pick } from 'ramda';
+
+import { getData } from '@centreon/ui';
 
 import {
   buildResourcesEndpoint,
@@ -12,14 +14,10 @@ import {
 } from './endpoint';
 import { ResourceListing, User, Resource } from '../models';
 
-const getData = <TData>({ endpoint, requestParams }): Promise<TData> =>
-  axios.get(endpoint, requestParams).then(({ data }) => data);
-
-const listResources = (
+const listResources = (cancelToken) => (
   endpointParams,
-  requestParams: AxiosRequestConfig = {},
 ): Promise<ResourceListing> =>
-  getData({ endpoint: buildResourcesEndpoint(endpointParams), requestParams });
+  getData<ResourceListing>(cancelToken)(buildResourcesEndpoint(endpointParams));
 
 interface AcknowledgeParams {
   acknowledgeAttachedResources?: boolean;
@@ -122,8 +120,8 @@ const checkResources = ({
   );
 };
 
-const getUser = (cancelToken) => (): Promise<User> =>
-  getData({ endpoint: userEndpoint, requestParams: cancelToken });
+const getUser = (token) => (): Promise<User> =>
+  getData<User>(token)(userEndpoint);
 
 export {
   acknowledgeResources,
