+++ www/front_src/src/Resources/index.tsx
@@ -1,30 +1,19 @@
 import * as React from 'react';
 
-import axios from 'axios';
-import { useSelector } from 'react-redux';
-import { isNil, equals } from 'ramda';
+import { isNil } from 'ramda';
 
-import { makeStyles, useTheme, Grid, Slide, fade } from '@material-ui/core';
+import { makeStyles, Slide } from '@material-ui/core';
 
-import { Listing, withSnackbar, useSnackbar, Severity } from '@centreon/ui';
+import { withSnackbar } from '@centreon/ui';
 
-import { listResources } from './api';
-import { ResourceListing, Resource, ResourceEndpoints } from './models';
-
-import { defaultSortField, defaultSortOrder, getColumns } from './columns';
+import { ResourceEndpoints } from './models';
+import Context from './Context';
 import Filter from './Filter';
-import ResourceActions from './Actions/Resource';
-import GlobalActions from './Actions/Refresh';
+import Listing from './Listing';
 import Details from './Details';
-import ApiNotFoundMessage from './ApiNotFoundMessage';
-import { rowColorConditions } from './colors';
-import { detailsTabId, graphTabId } from './Details/Body/tabs';
-import useFilter, { FilterState } from './Filter/useFilter';
-import {
-  labelSomethingWentWrong,
-  labelRowsPerPage,
-  labelOf,
-} from './translatedLabels';
+import useFilter from './Filter/useFilter';
+import useListing from './Listing/useListing';
+import useActions from './Actions/useActions';
 
 const useStyles = makeStyles((theme) => ({
   page: {
@@ -52,47 +41,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-type SortOrder = 'asc' | 'desc';
-
-const ResourceContext = React.createContext<FilterState | null>(null);
-
-const Resources = (): JSX.Element => {
-  const classes = useStyles();
-  const theme = useTheme();
-
-  const [listing, setListing] = React.useState<ResourceListing>();
-  const [selectedResources, setSelectedResources] = React.useState<
-    Array<Resource>
-  >([]);
-  const [resourcesToAcknowledge, setResourcesToAcknowledge] = React.useState<
-    Array<Resource>
-  >([]);
-  const [resourcesToSetDowntime, setResourcesToSetDowntime] = React.useState<
-    Array<Resource>
-  >([]);
-  const [resourcesToCheck, setResourcesToCheck] = React.useState<
-    Array<Resource>
-  >([]);
-
-  const [sorto, setSorto] = React.useState<SortOrder>(defaultSortOrder);
-  const [sortf, setSortf] = React.useState<string>(defaultSortField);
-  const [limit, setLimit] = React.useState<number>(30);
-  const [page, setPage] = React.useState<number>(1);
-
-  const listingContext = { listing, setListing };
-  const filterContext = useFilter();
-
-  const {
-    currentSearch,
-    setCurrentSearch,
-    nextSearch,
-    resourceTypes,
-    states,
-    statuses,
-    hostGroups,
-    serviceGroups,
-  } = filterContext;
-
+const useDetails = () => {
   const [
     selectedDetailsEndpoints,
     setSelectedDetailsEndpoints,
@@ -100,250 +49,46 @@ const Resources = (): JSX.Element => {
 
   const [detailsTabIdToOpen, setDefaultDetailsTabIdToOpen] = React.useState(0);
 
-  const [loading, setLoading] = React.useState(true);
-  const [enabledAutorefresh, setEnabledAutorefresh] = React.useState(true);
-
-  const refreshIntervalMs = useSelector(
-    (state) => state.intervals.AjaxTimeReloadMonitoring * 1000,
-  );
-  const refreshIntervalRef = React.useRef<number>();
-
-  const { showMessage } = useSnackbar();
-  const showError = (message): void =>
-    showMessage({ message, severity: Severity.error });
-
-  const [tokenSource] = React.useState(axios.CancelToken.source());
-
-  const load = (): void => {
-    setLoading(true);
-    const sort = sortf ? { [sortf]: sorto } : undefined;
-
-    listResources(
-      {
-        states: states.map(({ id }) => id),
-        statuses: statuses.map(({ id }) => id),
-        resourceTypes: resourceTypes.map(({ id }) => id),
-        hostGroupIds: hostGroups?.map(({ id }) => id),
-        serviceGroupIds: serviceGroups?.map(({ id }) => id),
-        sort,
-        limit,
-        page,
-        search: currentSearch,
-      },
-      { cancelToken: tokenSource.token },
-    )
-      .then((retrievedListing) => {
-        setListing(retrievedListing);
-      })
-      .catch((error) => {
-        setListing(undefined);
-
-        // if 404 is returned, it is probably an issue in apache configuration file
-        if (error.response?.status === 404) {
-          showError(ApiNotFoundMessage);
-        } else {
-          showError(error.response?.data?.message || labelSomethingWentWrong);
-        }
-      })
-      .finally(() => setLoading(false));
+  return {
+    selectedDetailsEndpoints,
+    setSelectedDetailsEndpoints,
+    detailsTabIdToOpen,
+    setDefaultDetailsTabIdToOpen,
   };
+};
 
-  const initAutorefresh = (): void => {
-    window.clearInterval(refreshIntervalRef.current);
+const Resources = (): JSX.Element => {
+  const classes = useStyles();
 
-    const interval = enabledAutorefresh
-      ? window.setInterval(load, refreshIntervalMs)
-      : undefined;
+  const listingContext = useListing();
+  const filterContext = useFilter();
+  const detailsContext = useDetails();
+  const actionsContext = useActions();
 
-    refreshIntervalRef.current = interval;
-  };
-
-  const initAutorefreshAndLoad = (): void => {
-    initAutorefresh();
-    load();
-  };
-
-  React.useEffect(() => {
-    return (): void => {
-      tokenSource.cancel();
-      clearInterval(refreshIntervalRef.current);
-    };
-  }, []);
-
-  React.useEffect(() => {
-    if (currentSearch !== nextSearch) {
-      return;
-    }
-    initAutorefreshAndLoad();
-  }, [
-    sortf,
-    sorto,
-    page,
-    limit,
-    currentSearch,
-    states,
-    statuses,
-    resourceTypes,
-    hostGroups,
-    serviceGroups,
-  ]);
-
-  React.useEffect(() => {
-    setCurrentSearch(nextSearch);
-  }, [states, statuses, resourceTypes, hostGroups, serviceGroups]);
-
-  React.useEffect(() => {
-    initAutorefresh();
-  }, [enabledAutorefresh]);
-
-  const changeSort = ({ order, orderBy }): void => {
-    setSortf(orderBy);
-    setSorto(order);
-  };
-
-  const changeLimit = (event): void => {
-    setLimit(Number(event.target.value));
-  };
-
-  const changePage = (_, updatedPage): void => {
-    setPage(updatedPage + 1);
-  };
-
-  const selectResources = (resources): void => {
-    setSelectedResources(resources);
-  };
-
-  const confirmAction = (): void => {
-    selectResources([]);
-    setResourcesToAcknowledge([]);
-    setResourcesToSetDowntime([]);
-    setResourcesToCheck([]);
-  };
-
-  const prepareToAcknowledge = (resources): void => {
-    setResourcesToAcknowledge(resources);
-  };
-
-  const prepareSelectedToAcknowledge = (): void => {
-    prepareToAcknowledge(selectedResources);
-  };
-
-  const cancelAcknowledge = (): void => {
-    prepareToAcknowledge([]);
-  };
-
-  const prepareToSetDowntime = (resources): void => {
-    setResourcesToSetDowntime(resources);
-  };
-
-  const prepareSelectedToSetDowntime = (): void => {
-    prepareToSetDowntime(selectedResources);
-  };
-
-  const cancelSetDowntime = (): void => {
-    prepareToSetDowntime([]);
-  };
-
-  const prepareToCheck = (resources): void => {
-    setResourcesToCheck(resources);
-  };
-
-  const prepareSelectedToCheck = (): void => {
-    prepareToCheck(selectedResources);
-  };
-
-  const columns = getColumns({
-    onAcknowledge: (resource) => {
-      prepareToAcknowledge([resource]);
-    },
-    onDowntime: (resource) => {
-      prepareToSetDowntime([resource]);
-    },
-    onCheck: (resource) => {
-      prepareToCheck([resource]);
-    },
-    onDisplayGraph: ({
-      details_endpoint,
-      status_graph_endpoint,
-      performance_graph_endpoint,
-    }) => {
-      setDefaultDetailsTabIdToOpen(graphTabId);
-      setSelectedDetailsEndpoints({
-        details: details_endpoint,
-        statusGraph: status_graph_endpoint,
-        performanceGraph: performance_graph_endpoint,
-      });
-    },
-  });
-
-  const selectResource = ({
-    details_endpoint,
-    status_graph_endpoint,
-    performance_graph_endpoint,
-  }): void => {
-    if (isNil(performance_graph_endpoint)) {
-      setDefaultDetailsTabIdToOpen(detailsTabId);
-    }
-    setSelectedDetailsEndpoints({
-      details: details_endpoint,
-      statusGraph: status_graph_endpoint,
-      performanceGraph: performance_graph_endpoint,
-    });
-  };
+  const {
+    detailsTabIdToOpen,
+    setDefaultDetailsTabIdToOpen,
+    selectedDetailsEndpoints,
+    setSelectedDetailsEndpoints,
+  } = detailsContext;
 
   const clearSelectedResource = (): void => {
     setSelectedDetailsEndpoints(null);
-  };
-
-  const toggleAutorefresh = (): void => {
-    setEnabledAutorefresh(!enabledAutorefresh);
   };
 
   const selectDetailsTabToOpen = (id): void => {
     setDefaultDetailsTabIdToOpen(id);
   };
 
-  const resourceDetailsOpenCondition = {
-    name: 'detailsOpen',
-    condition: ({ details_endpoint }): boolean =>
-      equals(details_endpoint, selectedDetailsEndpoints?.details),
-    color: fade(theme.palette.primary.main, 0.08),
-  };
-
-  const hasSelectedResources = selectedResources.length > 0;
-
-  const Actions = (
-    <Grid container>
-      <Grid item>
-        <ResourceActions
-          disabled={!hasSelectedResources}
-          resourcesToAcknowledge={resourcesToAcknowledge}
-          onPrepareToAcknowledge={prepareSelectedToAcknowledge}
-          onCancelAcknowledge={cancelAcknowledge}
-          resourcesToSetDowntime={resourcesToSetDowntime}
-          onPrepareToSetDowntime={prepareSelectedToSetDowntime}
-          onCancelSetDowntime={cancelSetDowntime}
-          resourcesToCheck={resourcesToCheck}
-          onPrepareToCheck={prepareSelectedToCheck}
-          onSuccess={confirmAction}
-        />
-      </Grid>
-      <Grid item style={{ paddingLeft: theme.spacing(3) }}>
-        <GlobalActions
-          disabledRefresh={loading}
-          enabledAutorefresh={enabledAutorefresh}
-          onRefresh={initAutorefreshAndLoad}
-          toggleAutorefresh={toggleAutorefresh}
-        />
-      </Grid>
-    </Grid>
-  );
-
-  const labelDisplayedRows = ({ from, to, count }): string =>
-    `${from}-${to} ${labelOf} ${count}`;
-
   return (
-    <ResourceContext.Provider value={{ ...listingContext, ...filterContext }}>
+    <Context.Provider
+      value={{
+        ...listingContext,
+        ...filterContext,
+        ...detailsContext,
+        ...actionsContext,
+      }}
+    >
       <div className={classes.page}>
         <div className={classes.filter}>
           <Filter />
@@ -369,37 +114,12 @@ const Resources = (): JSX.Element => {
             </Slide>
           )}
           <div className={classes.listing}>
-            <Listing
-              checkable
-              Actions={Actions}
-              loading={loading}
-              columnConfiguration={columns}
-              tableData={listing?.result}
-              currentPage={page - 1}
-              rowColorConditions={[
-                ...rowColorConditions(theme),
-                resourceDetailsOpenCondition,
-              ]}
-              limit={listing?.meta.limit}
-              onSort={changeSort}
-              onPaginationLimitChanged={changeLimit}
-              onPaginate={changePage}
-              sortf={sortf}
-              sorto={sorto}
-              labelRowsPerPage={labelRowsPerPage}
-              labelDisplayedRows={labelDisplayedRows}
-              totalRows={listing?.meta.total}
-              onSelectRows={selectResources}
-              selectedRows={selectedResources}
-              onRowClick={selectResource}
-              innerScrollDisabled={false}
-            />
+            <Listing />
           </div>
         </div>
       </div>
-    </ResourceContext.Provider>
+    </Context.Provider>
   );
 };
 
 export default withSnackbar(Resources);
-export { ResourceContext };
